Fix the 'english text on a dutch item'.
The error is caused because upload.php uses $upload->item to show the
parent's title and text. However, the parent ($item) is not the same
object as $upload->item. This means that a call to
$item->changeLocalization(true) (which will put the Dutch texts in
place), will change $item and $item->itemuploads[N] to Dutch, but not
$item->itemuploads[N]->item...

This is the reason that the News page didn't show the dutch texts in the
images caption (which is controlled by upload.php).

This commit includes a dirty hack. It's a shame, but for now it will
have to do...