Added hardcoded namespace constants for easy access.
The namespace model gives access to the namespaces in the database. In
the controllers (and other code), these namespaces currently referred to
using their ids (like 3 for project and 4 for graduation). This doesn't
help the readability.

So, now the ns model has 5 named constants that refer to the currently
existing namespaces: the other code can now refer to a namespace using
ns.PROJECT or ns.GRADUATION. I know this isn't very dynamic and future
proof, but so isn't the current code using the plain ids. Namespaces are
not added every week, so we're pretty safe for the moment.